Merge and split detections before and after track
Merge detections of consecutive videos to create tracks over all videos.
Split detections and tracks afterwards into parts to remain the file
structure. Every input file should have an output file with the same
name and the file ending .ottrk.
As metadata is no longer available after merging, the metadata is stored
and retrieved per frame. Normalization and denormalization have to be
changed to read and update the information correctly.