Pass log file path instead of log directory to OTVision logger
User has to specify log file path. If no path is specified
the logs will be saved in '{current working directory}/logs'
folder. The name of the log file will be the current datetime.
The overwrite flag must be set to `True` to overwrite a
pre-existing log file. Otherwise, an error will be raised.