Explicitly specify the venv when installing dependencies via uv
When OTVision is used as a third-party package, as is done in OTCloud, there are installation issues.
Instead of the install script installing the dependencies in OTVision's .venv, they are installed by uv in OTCloud.
By explicitly specifying the virtual environment, the OTVision dependencies are now installed in the correct environment.